requirements.txt
scrapy
requests
psycopg2-binary
supabase
python-dotenv